Capture service cmd before using in a 'here file'

+++ service/sync_gateway_service_install.sh
@@ -132,45 +132,47 @@ while [ "$1" != "" ]; do
     shift
 done
 
-# Check that runtime user account exists
-if [ "$OS" != "Darwin" ] && [ -z `id -u $RUNAS_TEMPLATE_VAR 2>/dev/null` ]; then
-    echo "The sync_gateway runtime user account does not exist \"$RUNAS_TEMPLATE_VAR\"." > /dev/stderr
-    exit 1
-fi
+if [ "$SERVICE_CMD_ONLY" = false ]; then
+    # Check that runtime user account exists
+    if [ "$OS" != "Darwin" ] && [ -z `id -u $RUNAS_TEMPLATE_VAR 2>/dev/null` ]; then
+        echo "The sync_gateway runtime user account does not exist \"$RUNAS_TEMPLATE_VAR\"." > /dev/stderr
+        exit 1
+    fi
 
-# Check that the runtime base directory exists
-if [ ! -d "$RUNBASE_TEMPLATE_VAR" ]; then
-    echo "The runtime base directory does not exist \"$RUNBASE_TEMPLATE_VAR\"." > /dev/stderr
-    exit 1
-fi
+    # Check that the runtime base directory exists
+    if [ ! -d "$RUNBASE_TEMPLATE_VAR" ]; then
+        echo "The runtime base directory does not exist \"$RUNBASE_TEMPLATE_VAR\"." > /dev/stderr
+        exit 1
+    fi
 
-# Check that the sync_gateway executable exists
-if [ ! -x "$GATEWAY_TEMPLATE_VAR" ]; then
-    echo "The sync_gateway executable does not exist \"$GATEWAY_TEMPLATE_VAR\"." > /dev/stderr
-    exit 1
-fi
+    # Check that the sync_gateway executable exists
+    if [ ! -x "$GATEWAY_TEMPLATE_VAR" ]; then
+        echo "The sync_gateway executable does not exist \"$GATEWAY_TEMPLATE_VAR\"." > /dev/stderr
+        exit 1
+    fi
 
-# Check that the sync_gateway src JSON config directory exists
-if [ ! -d "$SRCCFGDIR" ]; then
-    echo "The sync_gateway source JSON config file directory does not exist \"$SRCCFGDIR\"." > /dev/stderr
-    exit 1
-fi
+    # Check that the sync_gateway src JSON config directory exists
+    if [ ! -d "$SRCCFGDIR" ]; then
+        echo "The sync_gateway source JSON config file directory does not exist \"$SRCCFGDIR\"." > /dev/stderr
+        exit 1
+    fi
 
-# Check that the sync_gateway src JSON config file exists
-if [ ! -r "$SRCCFGDIR/$SRCCFG" ]; then
-    echo "The sync_gateway source JSON config file does not exist\"$SRCCFGDIR/$SRCCFG\"." > /dev/stderr
-    exit 1
-fi
+    # Check that the sync_gateway src JSON config file exists
+    if [ ! -r "$SRCCFGDIR/$SRCCFG" ]; then
+        echo "The sync_gateway source JSON config file does not exist\"$SRCCFGDIR/$SRCCFG\"." > /dev/stderr
+        exit 1
+    fi
 
-# If a /tmp/log_upr_client.sock socket exists from a previous installation remove it
-if [ -S /tmp/log_upr_client.sock ]; then
-    rm -f /tmp/log_upr_client.sock
-fi
+    # If a /tmp/log_upr_client.sock socket exists from a previous installation remove it
+    if [ -S /tmp/log_upr_client.sock ]; then
+        rm -f /tmp/log_upr_client.sock
+    fi
 
-# Copy a default config if defined config file does not exist
-if [ ! -e "$CONFIG_TEMPLATE_VAR" ]; then
-    cp $SRCCFGDIR/$SRCCFG $CONFIG_TEMPLATE_VAR
-    chown ${RUNAS_TEMPLATE_VAR}:${RUNAS_TEMPLATE_VAR} ${CONFIG_TEMPLATE_VAR}
+    # Copy a default config if defined config file does not exist
+    if [ ! -e "$CONFIG_TEMPLATE_VAR" ]; then
+        cp $SRCCFGDIR/$SRCCFG $CONFIG_TEMPLATE_VAR
+        chown ${RUNAS_TEMPLATE_VAR}:${RUNAS_TEMPLATE_VAR} ${CONFIG_TEMPLATE_VAR}
+    fi
 fi
 
 #Install the service for the specific platform
